+++ cmake/sources/kernels/cpu.cmake
@@ -27,7 +27,6 @@ SET(CPU_INCLUDE_DIR "${READDY_GLOBAL_DIR}/kernels/cpu/include")
 LIST(APPEND CPU_SOURCES "${SOURCES_DIR}/CPUKernel.cpp")
 LIST(APPEND CPU_SOURCES "${SOURCES_DIR}/CPUStateModel.cpp")
 LIST(APPEND CPU_SOURCES "${SOURCES_DIR}/data/DefaultDataContainer.cpp")
-LIST(APPEND CPU_SOURCES "${SOURCES_DIR}/data/NLDataContainer.cpp")
 LIST(APPEND CPU_SOURCES "${SOURCES_DIR}/observables/CPUObservableFactory.cpp")
 LIST(APPEND CPU_SOURCES "${SOURCES_DIR}/observables/CPUObservables.cpp")
 
@@ -43,4 +42,5 @@ LIST(APPEND CPU_SOURCES "${SOURCES_DIR}/actions/reactions/ReactionUtils.cpp")
 LIST(APPEND CPU_SOURCES "${SOURCES_DIR}/actions/reactions/Event.cpp")
 LIST(APPEND CPU_SOURCES "${SOURCES_DIR}/actions/reactions/CPUUncontrolledApproximation.cpp")
 LIST(APPEND CPU_SOURCES "${SOURCES_DIR}/actions/reactions/CPUGillespie.cpp")
+LIST(APPEND CPU_SOURCES "${SOURCES_DIR}/actions/topologies/CPUTopologyActions.cpp")
 LIST(APPEND CPU_SOURCES "${SOURCES_DIR}/actions/topologies/CPUTopologyActionFactory.cpp")
