+++ kernels/cpu/include/readdy/kernel/cpu/observables/CPUObservables.h
@@ -31,7 +31,6 @@
 
 #pragma once
 #include <readdy/model/observables/Observables.h>
-#include <readdy/kernel/cpu/data/NLDataContainer.h>
 
 namespace readdy {
 namespace kernel {
