+++ kernels/cpu/test/TestReactions.cpp
@@ -67,7 +67,7 @@ TEST(CPUTestReactions, CheckInOutTypesAndPositions) {
     using conversion_t = readdy::model::reactions::Conversion;
     using death_t = readdy::model::reactions::Decay;
     using particle_t = readdy::model::Particle;
-    using data_t = readdy::kernel::cpu::data::NLDataContainer;
+    using data_t = readdy::kernel::cpu::data::DefaultDataContainer;
     auto kernel = std::make_unique<readdy::kernel::cpu::CPUKernel>();
     kernel->context().periodicBoundaryConditions() = {{false, false, false}};
     kernel->context().boxSize() = {{100, 100, 100}};
